// src/types/hr-management/employee-management.d.ts
export interface ISelect {
    minGrade?: number;
    maxGrade?: number;
    id?: string;
    label?: number | string;
    value?: number | string;
}

export interface IEntryLanguage {
    entryLanguageName: string;
}

export interface IEditInformationProps {
    isShow: boolean;
    onCancel: () => void;
    data: IData | undefined;
    reloadAPIEmployee?: (params: object) => void;
    editedFields?: string[];
    isLimit: boolean;
    moduleName?: string;
}

export interface IEditInformationForm {
    lastName: string;
    firstName: string;
    genderId: number;
    birthday: Date;
    maritalStatusId: number;
    birthPlace: string;
    personalEmail: string;
    mobilePhone: string;
    homePhone: string;
    nationalityId: number;
    isDegree: boolean;
    isReceivedIdCard: boolean;
    isReceivedCV: boolean;
    isTranscript: boolean;
    isHouseholdRegistration: boolean;
}
// ...
export interface IMaritalStatus {
    maritalStatusId: number;
    maritalStatusName: string;
}